remove divide and project commit

+++ application/admin/model/HunningtuModel.php
@@ -77,10 +77,13 @@ class HunningtuModel extends Model
     }
 
     public function delHunningtuByUid($uid){
-        $bol = $this->where('uid',$uid)->delete();
-        if($bol){
-            return ['code' => 1, 'data' => '', 'msg' => '混凝土删除成功'];
+        $has = $this->where('uid',$uid)->value('id');
+        if($has){
+            $bol = $this->where('uid',$uid)->delete();
+            if($bol < 1){
+                return ['code' => 1, 'data' => '', 'msg' => '混凝土删除失败'];
+            }
         }
-        return ['code' => 0, 'data' => '', 'msg' => '混凝土删除失败'];
+        return ['code' => 1, 'data' => '', 'msg' => '混凝土删除成功'];
     }
 }
